lib: refactoo move effects + Crashfx to SCLang module
