--- renardo_lib/renardo_lib/Effects/__init__.py
@@ -1,1 +1,0 @@
-from renardo_lib.Effects.Util import *
