+++ renardo_lib/renardo_lib/Players.py
@@ -122,11 +122,11 @@
 
 """
 
-from copy import copy, deepcopy
+from copy import copy
 
 from renardo_lib.Settings import SamplePlayer, LoopPlayer, SYNTHDEF_DIR
 from renardo_lib.SCLang.SynthDef import SynthDefProxy, SynthDef, SynthDefs
-from renardo_lib.Effects import FxList
+from renardo_lib.SCLang.Effects import FxList
 from renardo_lib.Buffers import Samples
 from renardo_lib.Key import PlayerKey, NumberKey
 from renardo_lib.Repeat import Repeatable
