+++ renardo_lib/renardo_lib/SCLang/Effects/crashFX.py
@@ -1,5 +1,3 @@
-from ..Effects import *
-
 # Legato slide
 fx = FxList.new("leg", "leg", {"leg":0, "sus":1 }, order = 0)
 fx.add("osc = osc * XLine.ar(Rand(0.5,1.5)*leg,1,0.05*sus)")
